Converted examples from double --> float
git-svn-id: https://aquitaine.dnsalias.net:8443/svn/mathlib2/trunk@9 35ab40ff-a1d5-ea4a-9275-db5122b6521a

+++ CatmullRomSplines/CatmullRomSplines.cs
@@ -41,7 +41,7 @@ namespace CatmullRomSplines
 			const float width = 6f;
 
 			// Create point bounds
-			RectangleF pointBounds = new RectangleF(point.X - 0.5f * width, point.Y - 0.5f * width, width, width);
+			RectangleF pointBounds = new RectangleF((float)point.X - 0.5f * width, (float)point.Y - 0.5f * width, width, width);
 
 			// Create Pen
 			Brush brush = new SolidBrush(color);
@@ -60,8 +60,8 @@ namespace CatmullRomSplines
 			Pen pen = new Pen(color, width);
 
 			// Create Drawing.Points
-			PointF pt1 = new PointF(point1.X, point1.Y);
-			PointF pt2 = new PointF(point2.X, point2.Y);
+			PointF pt1 = new PointF((float)point1.X, (float)point1.Y);
+			PointF pt2 = new PointF((float)point2.X, (float)point2.Y);
 
 			// Draw point
 			graphics.DrawLine(pen, pt1, pt2);
